requirements.txt
distlib==0.3.4
filelock==3.4.2
platformdirs==2.4.1
PyJWT==2.3.0
six==1.16.0
virtualenv==20.13.0
